+++ src/lib/hooks/useTimer/index.ts
