+++ src/lib/hooks/useTimer/lib/timerWorker.ts
