+++ src/lib/hooks/useTimer/model/TimerMessageOut.ts
