+++ src/lib/hooks/useTimer/model/consts.ts
