+++ src/lib/snippets/PickExtends.ts
