+++ src/lib/snippets/fromSeconds.ts
