+++ src/model/types/TimerSettings/TimerSettings.ts
