+++ src/model/types/TimerSettings/index.ts
