+++ src/res/style/styles.ts
