+++ src/ui/pages/Home/model/HomePageStatic.ts
