+++ src/ui/pages/Home/model/svg.ts
