+++ src/ui/pages/Home/ui/BackgroundOverlay.tsx
