+++ src/ui/pages/Home/ui/TimerInput.tsx
