+++ tsconfig.json
@@ -23,6 +23,6 @@
       "@Parser/*": ["./src/domains/Parser/*"]
     }
   },
-  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts", "src/snippets/useOnUnMount.tm"],
+  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts", "src/lib/snippets/useOnUnMount.tm"],
   "exclude": ["node_modules"]
 }
